backend/requirements.txt
Flask==2.2.5
requests==2.31.0
flask-cors==3.0.10
python-dotenv